requirement.txt
streamlit
pillow
pytesseract
google-generativeai
python-dotenv
